# api-spec.yaml
openapi: 3.0.1
info:
  title: CirendsAPI
  version: "1.0"
servers:
  - url: https://cirends.runasp.net/api
  - url: http://localhost:5000/api
...
security:
  - Bearer: []
components:
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header
      description: "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
...
paths:
...
  /activities/{activityId}/expenses/{expenseId}/shares/{shareId}/mark-paid:
    patch:
      tags: [ExpenseShares]
      parameters:
        - { name: activityId, in: path, required: true, schema: { type: integer } }
        - { name: expenseId, in: path, required: true, schema: { type: integer } }
        - { name: shareId, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: No Content }
        "400": { description: Bad Request }
        "403": { description: Forbidden }
        "404": { description: Not Found }
  /activities/{activityId}/expenses/{expenseId}/shares/{shareId}/unmark-paid:
    patch:
      tags: [ExpenseShares]
      parameters:
        - { name: activityId, in: path, required: true, schema: { type: integer } }
        - { name: expenseId, in: path, required: true, schema: { type: integer } }
        - { name: shareId, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: No Content }
        "400": { description: Bad Request }
        "403": { description: Forbidden }
        "404": { description: Not Found }
  /activities/{activityId}/expenses/mark-all-paid:
    post:
      tags: [ExpenseShares]
      parameters:
        - { name: activityId, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: No Content }
        "400": { description: Bad Request }
        "403": { description: Forbidden }
        "404": { description: Not Found }
...